Buscas e validação das ações inserir e alterar

+++ src/main/java/com/projeto/model/models/Departamento.java
@@ -14,7 +14,7 @@ import javax.persistence.Table;
 @Table(name = "TAB_DEPARTAMENTO")
 public class Departamento {
 	
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	private List<Usuario> usuario;
@@ -23,7 +23,7 @@ public class Departamento {
 		super();
 	}
 	
-	public Departamento(Long id, String nome) {
+	public Departamento(Integer id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -33,11 +33,11 @@ public class Departamento {
 	@Id
 	@GeneratedValue(strategy= GenerationType.IDENTITY)
 	@Column(name="id", nullable = false)
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
